+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.2.0
+- optional cryptographic validation of incoming webhooks (-validate option)
+- server runs either standard or ssl, not both
+
 ## v0.1.0
 - support encrypted websockets